requirements.txt
python-multipart==0.0.5
colorgram.py==1.2.0
fastapi==0.67.0
aiopg==1.2.1
SQLAlchemy~=1.3.19
